# improve performance of `total_mempool_fees()` and `total_mempool_cost()`