### Purpose:

Currently, the mempool asks sqlite what the total cost and total fees are of the items in the mempool. This is very reliable as all the book-keeping is done by sqlite. However, it's also slow. It appears sqlite does not maintain these sums continuously, but actually computes them every time we ask.

This patch adds "manual" book-keeping of the total cost and total fees in the mempool, avoiding a SQL query every time we want to know.

### Current Behavior:

`total_mempool_fees()` and `total_mempool_cost()` are expensive calls.

### New Behavior:

`total_mempool_fees()` and `total_mempool_cost()` are cheap calls.


### profiles

| benchmark | before | after | after/before |
| --- | --- | --- | --- |
| add_spend_bundle() | 106.4s | 58.5s | 55% |
| add_spend_bundle() with replace-by-fee | 198.1s | 60.9s | 30.76% |

#### Before
![mempool-add-st](https://github.com/Chia-Network/chia-blockchain/assets/661450/215d78bd-fbf0-4d1d-bbd1-7d4fca3137bb)

#### After
![mempool-add-st](https://github.com/Chia-Network/chia-blockchain/assets/661450/4ac80fea-c1a2-42b5-9fb4-1c426de89588)