@neurosis69 2023-12-20
It may be worth to run the profiler with these 3 pragma changes as well, if you haven't tried them.

[__cache_size__](https://www.sqlite.org/pragma.html#pragma_cache_size)
__What__: &emsp;Change sqlites cache size from the default 2 MiB to 200 MiB
__Why__:  &ensp;&emsp;Suggested a random number just to find out if a bigger sqlite cache would improve code. The required size can than probably be calculated.

[_journal_mode_](https://www.sqlite.org/pragma.html#pragma_journal_mode)
__What__: &emsp;Change sqlites journal_mode from the default MEMORY to OFF.
__Why__: &ensp;&emsp;I didn't find any explicit rollbacks/commits in the code and my understanding of this mempool is, that it gets feed from the peers as soon as it is started up -> likely no journal required?

[_temp_store_](https://www.sqlite.org/pragma.html#pragma_temp_store)
__What__: &emsp;Change sqlites temp_store from the default 0, which means file, to 2, which means memory.
__Why__: &ensp;&emsp;My understanding is, that in theory even if an in-memory sqlite database is used and as long as this parameter is set to 0, the indexes aren't stored in memory. Doesn't sound reasonable but probably worth to verify if it has an impact imho.

```sql
    def __init__(self, mempool_info: MempoolInfo, fee_estimator: FeeEstimatorInterface):
        self._db_conn = sqlite3.connect(":memory:")
        self._items = {}
        self._block_height = uint32(0)
        self._timestamp = uint64(0)

        with self._db_conn:
            self._db_conn.execute("PRAGMA cache_size = -200000;")
            self._db_conn.execute("PRAGMA journal_mode = OFF;")
            self._db_conn.execute("PRAGMA temp_store = 2;")
```