@arvidn 2023-12-20
@neurosis69 the mempool's sqlite database is `":memory:"` already. I wouldn't expect any of those pragmas to make a difference