@arvidn 2023-12-20
my understanding is that a `:memory:` database can only have a single connection, so it's already "single threaded" so to speak. Since it's in memory, there's no fault tolerance required, no journaling. Would sqlite really enable such redundant features by default for an in-memory database?