@neurosis69 2023-12-20
I'm with you, it doesn't sound reasonable but there are still some hints in the documentation, that's why I asked.
Don't have much time myself to set up a testing environment quickly, but I thought you may be able to quickly execute some runs and verify it. 

For instance, this info is from the journal_mode parameter:
```
Note that the journal_mode for an [in-memory database](https://www.sqlite.org/inmemorydb.html) is either MEMORY or OFF and can not be changed to a different value. An attempt to change the journal_mode of an [in-memory database](https://www.sqlite.org/inmemorydb.html) to any setting other than MEMORY or OFF is ignored. Note also that the journal_mode cannot be changed while a transaction is active.
```
If it's safe to deactivate it I'd guess that a lots of sqlite internal code is skipped.