@arvidn 2023-12-21
thanks for the review! I've updated it to move the accounting into `remove_from_pool()`, which should make it more reliable. Now I query the cost- and fee sums out of the table.